requirements.txt
# Core framework (use the official PyTorch index to grab CUDA-specific builds if required)
torch==2.3.1
torchvision==0.18.1

# Ultralytics (editable install tracked separately)
-e git+https://github.com/ultralytics/ultralytics.git@8da29ee37a7d7bd8e8d2a02d2fe706d319c2da62#egg=ultralytics

# Ultralytics dependencies (versioned to match current environment)
matplotlib==3.10.7
numpy==2.1.2
opencv-python==4.12.0.88
polars==1.34.0
polars-runtime-32==1.34.0
psutil==7.1.1
PyYAML==6.0.1
requests==2.25.1
scipy==1.15.3
ultralytics-thop==2.0.17
